# stringslice: use strings.EqualFold for HasI