Hi.

This PR uses `strings.EqualFold` instead of  `strings.ToLower(current) == strings.ToLower(needle)` in `stringslice.HasI` for case-insensitive comparison. This change make `HasI` almost 3x faster.

```
benchmark            old ns/op     new ns/op     delta
BenchmarkHasI-12     90.0          33.2          -63.11%

benchmark            old allocs     new allocs     delta
BenchmarkHasI-12     0              0              +0.00%

benchmark            old bytes     new bytes     delta
BenchmarkHasI-12     0             0             +0.00%
```